Support 5.0
The AST hasn't changed from 4.14 to 5.0. Furthermore, the magic numbers
from 4.14 haven't been bumped (yet). So the only thing that needs to be
taken care of is the dune `ocaml_version` variable.

IMPORTANT: This might only be a temporary solution: if the magic numbers
still get bumped, ppxlib will need to add some extra logic to support
those magic numbers.

Signed-off-by: Sonja Heinze <sonjaleaheinze@gmail.com>